+++ src/webapp/controllers/UsersController.php
@@ -123,14 +123,17 @@ class UsersController extends Controller
 
     public function destroy($username)
     {
-        if ($this->userRepository->deleteByUsername($username) === 1) {
-            $this->app->flash('info', "Sucessfully deleted '$username'");
-            $this->app->redirect('/admin');
-            return;
+        if($this->auth->isAdmin()){
+		if ($this->userRepository->deleteByUsername($username) === 1) {
+		    $this->app->flash('info', "Sucessfully deleted '$username'");
+		    $this->app->redirect('/admin');
+		    return;
+		}      
+		$this->app->flash('info', "An error ocurred. Unable to delete user '$username'.");
+		$this->app->redirect('/admin');
+        }else{
+              $this->app->redirect('/admin');
         }
-
-        $this->app->flash('info', "An error ocurred. Unable to delete user '$username'.");
-        $this->app->redirect('/admin');
     }
 
     public function makeSureUserIsAuthenticated()
